switching it back... bigger problem here

+++ source/java/fds-project/fds/src/main/java/com/formationds/client/v08/converters/ExternalModelConverter.java
@@ -491,7 +491,7 @@ public class ExternalModelConverter {
 			VolumeSettingsObject objectSettings = (VolumeSettingsObject) externalVolume.getSettings();
 
 			if ( objectSettings.getMaxObjectSize() != null ){
-				internalSettings.setMaxObjectSizeInBytes( objectSettings.getMaxObjectSize().getValue( SizeUnit.B ).longValue() );
+				internalSettings.setMaxObjectSizeInBytes( objectSettings.getMaxObjectSize().getValue( SizeUnit.B ).intValue() );
 			}
 			else {
 				internalSettings.setMaxObjectSizeInBytes( DEF_OBJECT_SIZE );
